% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualResume_function.R
\name{Visual.Resume}
\alias{Visual.Resume}
\title{Creates a visual resume}
\usage{
Visual.Resume(titles.left = c("Main Title", "Sub-title", "Sub-Sub-title"),
  titles.left.cex = c(3, 2, 1), titles.right = c("www.a.com",
  "me@gmail.com", ""), titles.right.cex = c(3, 2, 1), center.labels = c("D",
  "E"), top = data.frame(title = c("College A", "University B",
  "University C"), sub = c("a"), start = c(2001.5, 2006.5, 2011), end =
  c(2005.5, 2010.9, 2014.5)), bottom = data.frame(title = c("Project A",
  "Project B", "Project C"), sub = c("b"), start = c(2010, 2014.6, 2014.6), end
  = c(2010.5, 2015.2, 2015.2)), milestones = data.frame(title = c("BA", "MS",
  "Ph.D"), subtitle = c("2005", "2010", "2014"), date = c(2005.5, 2010.9,
  2014.5)), events = data.frame(year = c(2007, 2011.5, 2014.2), title = c("a",
  "b", "c")), events.cex = 1, year.steps = 1, interests = list(Programming
  = c("R", "Javascript", "HTML"), Statistics = c("Bayesian", "Regression",
  "Decision\\nTrees"), Research = c("Decision Making", "Foraging")),
  font.family = NA, col = "basel", trans = 0.6, year.range = NULL)
}
\arguments{
\item{titles.left, titles.right}{character. Vector of up to length three indicating labels to be shown in the top left / right.}

\item{titles.left.cex, titles.right.cex}{numeric. Vector indicating the size of the respective labels.}

\item{center.labels}{character. Vector of two labels for the top and bottom sections of the timeline.}

\item{top}{dataframe. Specifications of the top section of the timeline.}

\item{bottom}{dataframe. Specificaions of the bottom section of the timeline.}

\item{milestones}{dataframe. Specifications of brief level-1 timeline milestones (at top of timeline).}

\item{events}{dataframe. Specificaions of longer level-2 milestones (in bottom right plot).}

\item{events.cex}{numeric. A size multiplier for level-2 milestones}

\item{year.steps}{integer.}

\item{interests}{list. A list of length up to 4 indicating one's interests / skills. Each entry should be a character vector up to length 5. The more often a value occurs in the vector, the more pronounced it will be.}

\item{font.family}{character. An optional google font family. Run \code{sysfonts::font.families.google()} to see all of them. Only use when creating plot in a plotting device such as pdf()}

\item{col}{character. The color palette for the plot. Can be a named palette from the \code{yarrr::piratepal()} function, or a vector of named colors.}

\item{trans}{numeric. A number between 0 and 1 indicating the transparencies of the colors}

\item{year.range}{integer. An optional vector of minimum and maximum years to plot in the timeline.}
}
\description{
This function takes several arguments and returns a visual resume plot. The plot is optimized for an 8.5 in x 11 in display.
}
\details{
The middle, timeline in the plot is plotted with y values ranging from 0 to 100.
 The arguments \code{top} and \code{bottom}, specifying events in the bottom and top of the timeline should each be a dataframe with at least the columns title, sub, start, and end. To specify plotting locations explicitly, you can include the columns \code{box.x0, box.x1, box.y0, box.y1} for the box locations and/or \code{label.x, label.y, label.dir} for the label coordinates and direction ("left" or "right")
}
\examples{






Visual.Resume(
 titles.left = c("Walter White, PhD", "Chemistry, Cooking, Pizza", "*Built with love in R using the InfoResume package: www.ndphillips.github.io/inforesume"),
 titles.right = c("www.lospolloshermanos.com", "TheOneWhoKnocks@gmail.com", "Full Resume: www.ndphillips.github.io"),
 titles.right.cex = c(2, 2, 1),
 titles.left.cex = c(4, 2, 1),
 center.labels = c("Education", "Teaching"),
 top = data.frame(title = c("Grinnell College", "Ohio U", "U of Basel", "MPIB", "U of Konstanz", "U of Basel"),
                  sub = c("BA. Student", "MS. Student", "PhD. Student", "PhD. Researcher", "PostDoc", "PostDoc"),
                  start = c(1976, 1980, 1982, 2012, 2014.7, 2016.1),
                  end = c(1980, 1982, 1985, 2014.6, 2016, 2017)),
 bottom = data.frame(title = c("Musician's Friend", "Statistics", "Research Methods", "Information Search", "R", "R", "R", "R"),
                     sub = c("Statistician", "OhioU", "OhioU", "Konstanz", "Konstanz", "Konstanz", "Basel", "Basel"),
                     start = c(2005.5, 2009.5, 2010, 2014.8, 2014.8, 2015.4, 2016.1, 2016.7),
                     end = c(2006.5, 2010, 2011, 2015.3, 2015.3, 2015.9, 2016.6, 2017)),
 milestones = data.frame(title = c("BA Mathematics", "MS Chemistry", "Ph.D Chemistry"),
                         subtitle = c("2005", "2010", "2014"),
                         date = c(2005.5, 2010.9, 2014.5)),
 events = data.frame(year = c(2016, 2015, 2014, 2013, 2012),
                     title = c("White, W. (2016). YaRrr! The Pirate's Guide to R.\\nSelf-published e-book.",
                               "White, W. & Pinkman, J. (2015). The Janus face of Darwinian competition.\\nScientific Reports, 5.",
                               "White, W. & Pinkman, J. (2014). Rivals in the Dark: How competition influences [...].\\nCognition, 113(1), 104-119.",
                               "White, W. & Pinkman, J. (2013). Early positive information impacts final [...].\\nJournal of Behavioral Decision Making, 27.",
                               "White, W. & Pinkman, J. (2012). Predicting soccer matches [...].\\nJudgment and Decision Making. 5(3). 200-206.")),
 interests = list("programming" = c(rep("R", 10), rep("Python", 1), rep("JavaScript", 2), "MatLab"),
                  "statistics" = c(rep("Decision Trees", 10), rep("Bayesian", 5), rep("Regression", 3)),
                  "research" = c(rep("Information Search", 10), rep("Decision Making", 5), rep("Statistical Reasoning", 3))
                  #            "test2" = c(rep("AAA", 10), rep("BBB", 50), rep("CCC", 3))
 ),
 col = "xmen",
 trans = .6)
}

